+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * **ProjectGuidsMustBeUniqueRule**: Verifies that every project in the solution has a unique GUID.
 * **NuGetReferenceHintPathsMustBeValidRule**: Verifies that all NuGet references in the project have correct hint paths (pointing to an actually existing file). 
 * **NonDevelopmentNuGetReferencesShouldHaveIsPrivateSetToTrueRule**: Verifies that all non-development NuGet references in the project have their 'IsPrivate' flag (also referred to as 'Copy Local') set to true. 
+* **ProjectItemMustBeIncludedByWildcardRule/ProjectItemMustNotBeIncludedByWildcardRule**: Verifies that project items are included/not included by wildcards.
 
 #### Changed
 * Renamed AllProjectItemsMustBePresentRule to ProjectItemsMustBePresentRule to be more in line with the newly added rules.